Upgrade to v0.11.0 - Reorganize server directories
