+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tw5-tickets",
-  "version": "0.10.10",
+  "version": "0.11.0",
   "description": "Search TiddlyWiki5 repository issues and pull requests",
   "main": "server.js",
   "scripts": {
