+++ server/sockets/get-tiddler.js
