+++ server/sockets/set-tiddler.js
